# lib/api-zod/src/generated/types/investigationSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APK Fraud Intelligence Investigator API
 * OpenAPI spec version: 0.1.0
 */

export interface InvestigationSummary {
  id: number;
  sampleName: string;
  sha256: string;
  /** @nullable */
  packageName?: string | null;
  verdict: string;
  riskLevel: string;
  riskScore: number;
  confidence: string;
  primaryThreatType: string;
  /** @nullable */
  clusterId?: string | null;
  createdAt: Date;
}