# lib/api-zod/src/generated/types/permissionAbuse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APK Fraud Intelligence Investigator API
 * OpenAPI spec version: 0.1.0
 */

export interface PermissionAbuse {
  permission: string;
  riskLevel: string;
  explanation: string;
}