# lib/api-zod/src/generated/types/riskMatrix.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APK Fraud Intelligence Investigator API
 * OpenAPI spec version: 0.1.0
 */

export interface RiskMatrix {
  severity: string;
  impact: number;
  likelihood: number;
  composite: number;
  methodology: string;
  /** @nullable */
  cvssVector?: string | null;
}